Props definition correction
